+++ Framework/Reference/Engine/include/NodeSet.h
@@ -16,14 +16,14 @@ public:
 
 	inline const Node* GetRootNode() const { return m_rootNode; }
 
-	inline size_t GetNodeCount() const { return m_nodesByIndex.size(); }
+	inline uint GetNodeCount() const { return uint(m_nodesByIndex.size()); }
 
 	inline const Node* GetNodeByIndex(uint index) const { return m_nodesByIndex[index]; }
 
 	const Node* GetNodeByName(const tstring& name) const;
 
 	_declspec(property(get = GetRootNode)) const Node* rootNode;
-	_declspec(property(get = GetNodeCount)) size_t nodeCount;
+	_declspec(property(get = GetNodeCount)) uint nodeCount;
 
 private:
